move api routes to pages
